resolve: modify the code
