+++ src/bk-user/tests/apps/tenant/test_tasks.py
@@ -16,7 +16,7 @@
 # to the current version of the project delivered to anyone in the future.
 import pytest
 from bkuser.apps.tenant.constants import TenantUserStatus
-from bkuser.apps.tenant.periodic_tasks import update_expired_tenant_user_status
+from bkuser.apps.tenant.tasks import update_expired_tenant_user_status
 
 pytestmark = pytest.mark.django_db
